requirements_dev.txt
-r requirements.txt
ipython==4.0.0
pytest-django==2.9.1
coverage==4.0.1